+++ tedi/components/helpers/scroll-fade/scroll-fade.stories.ts
@@ -8,7 +8,7 @@ const loremIpsum = `Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed
   consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.`;
 
 /**
- * <a href="https://www.figma.com/design/jWiRIXhHRxwVdMSimKX2FF/TEDI-READY-2.39.64?node-id=10758-111141&m=dev" target="_blank">Figma ↗</a><br>
+ * <a href="https://www.figma.com/design/jWiRIXhHRxwVdMSimKX2FF/TEDI-READY-2.39.64?node-id=10758-111141&m=dev" target="_blank">Figma ↗</a><br />
  * <a href="https://www.tedi.ee/1ee8444b7/p/32b155-scroll-fade" target="_blank">Zeroheight ↗</a>
  */