+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+# [7.1.0-rc.7](https://github.com/TEDI-Design-System/angular/compare/angular-7.1.0-rc.6...angular-7.1.0-rc.7) (2026-07-15)
+
+
+### Bug Fixes
+
+* **pagination:** community pagination results rendered twice [#548](https://github.com/TEDI-Design-System/angular/issues/548) ([2ba2bb3](https://github.com/TEDI-Design-System/angular/commit/2ba2bb36b1805dac18adf1c5c929962b7ec7146c))
+
+# [7.1.0-rc.6](https://github.com/TEDI-Design-System/angular/compare/angular-7.1.0-rc.5...angular-7.1.0-rc.6) (2026-07-14)
+
+
+### Bug Fixes
+
+* **dropdown:** skip close on certain events [#544](https://github.com/TEDI-Design-System/angular/issues/544) ([#545](https://github.com/TEDI-Design-System/angular/issues/545)) ([f9be659](https://github.com/TEDI-Design-System/angular/commit/f9be659821097949648c7896a77cc8d116af1169))
+
+# [7.1.0-rc.5](https://github.com/TEDI-Design-System/angular/compare/angular-7.1.0-rc.4...angular-7.1.0-rc.5) (2026-07-10)
+
+
+### Features
+
+* **breadcrumbs:** new TEDI-ready component [#512](https://github.com/TEDI-Design-System/angular/issues/512) ([#518](https://github.com/TEDI-Design-System/angular/issues/518)) ([6691cd2](https://github.com/TEDI-Design-System/angular/commit/6691cd25a1c93ad286b75a84e65fe212ae089457))
+
 # [7.1.0-rc.4](https://github.com/TEDI-Design-System/angular/compare/angular-7.1.0-rc.3...angular-7.1.0-rc.4) (2026-07-09)